requirements.txt
# --extra-index-url https://download.pytorch.org/whl/cu118
torch
torchvision
torchaudio
setuptools
wheel
pandas == 2.0.3
numpy >= 1.24.4
PyYAML == 6.0.1
python-box == 7.1.1
ensure == 1.0.3
tqdm == 4.66.1
types-PyYAML == 6.0.12.12
joblib == 1.3.2
pathlib == 1.0.1
transformers == 4.38.2
sentencepiece == 0.2.0
scikit-learn == 1.4.0
lightgbm == 4.3.0
